+++ src/pkgmt/new.py
@@ -12,7 +12,7 @@ def render_inplace(path, **kwargs):
     path.write_text(template.safe_substitute(**kwargs))
 
 
-def package(name):
+def package(name, use_setup_py=False):
     """Create a new package"""
     project_name = name.replace("_", "-")
     package_name = name.replace("-", "_")
@@ -32,6 +32,7 @@ def package(name):
         "tasks.py",
         "MANIFEST.in",
         "pyproject.toml",
+        "pyproject-setup.toml",
         ".github/workflows/ci.yml",
     ):
         render_inplace(
@@ -41,3 +42,12 @@ def package(name):
         )
 
     os.rename(root / "src" / "package_name", root / "src" / package_name)
+
+    if use_setup_py:
+        (root / "pyproject-setup.toml").unlink()
+    else:
+        (root / "pyproject.toml").unlink()
+        (root / "setup.py").unlink()
+        (root / "MANIFEST.in").unlink()
+        (root / "pyproject-setup.toml").rename(root / "pyproject.toml")
+        (root / "src" / package_name / "__init__.py").write_text("")